Split out useRenameActiveFile
